Working on the bar ordering step
Struggling through this one but I have made progress. I figured out how
make a button sort a data array of objects when clicked. I set up a
simple example but running into issues when testing in the console. I'm
confused about when the console and my app has access to certain
variables and when it doesn't.