Merge latest component library into my bar chart branch
I wanted play around with my bar chart in the new set up but I was getting error messages when running the app. I spent a long time troubleshooting and found that the issue was no areaCodeLookup in the testData file.
Now the app is running. However, now I realise that there are a host of other problems trying to play arounf with the bar chart component in the new set up.
I need to rework my component to make it fit with the new environment. e.g. file names, paths, etc.